Double Jump and faster move
There is a bug with the way jumping interacts with colliding with
blocks. (touching a block in any way reads as grounded, which allows
another jump). Should be fixable by adding an invisible plate on top of
each block for the character to collide with, instead of the actual
visible block