Slow block and Bounce block added
Implementation of bounce is just to literally bounce the player, this
felt more intuitive. It would be easy to implement a buff to the player
that gives him greater jump abilities, if that is what you all had in
mind.

Also, feel free to play with the public values I have defined for things
like speed up/speed down/bounce height etc. This was just a first pass
at the actual intensity of the blocks' effects.

Lastly, the current implementation is that the effects only occur when
the player collides with the top of the block (the plate). Down the road
it may be necessary to apply the effects when colliding from any
direction.