powerups and collectables
-Green and red power-ups do not affect blue blocks. I thought it was
worth preserving the ability to take a green block into a blue block for
a "long jump". Easy to change, should we decide otherwise.
-Powerups currently only affect other *colored* blocks. I tried having
all blocks affected by powerups, but it introduces a scenario where it
becomes pretty easy to speed through the entire level. I can switch this
back if you all prefer.
-Variables for the block effects are now located on the player and
static. I know this is annoying, but I just didn't have time to find a
way around this. Can revisit after midterms.

-Lastly, there is an issue with assigning rigidbodies to prefabs, such
that a prefab won't allow any assigning to a rigidbody that's in the
scene. Basically, what this means is that instead of using the prefab
blocks to build the level, you will need to: place an initial block from
the prefab, and assign the playerCylinder to the blocks Plate script in
the inspector. Once that block works, simply duplicate it to place
another block, rather than dragging another in from the prefabs. I can
show you exactly what I mean in detail when we meet. It isn't
complicated, and won't slow down level design, just difficult to explain
in text.

Also, the collectable cassettes are implemented. They increment a
variable that is located on the player script (points).